Modularize and get the project running (#31)
* Get the base project running

* updated readme
