--- requirements.txt
@@ -1,2 +1,0 @@
-pandas==0.23.4
-matplotlib==3.0.2
