+++ web/requirements.txt
@@ -1,0 +1,2 @@
+flask==1.1.2
+flask_sqlalchemy==2.4.1
