+++ web/setup.py
@@ -1,0 +1,8 @@
+from setuptools import setup
+
+setup(
+    name='TESS',
+    packages=['web'],
+    include_package_data=True,
+    install_requires=['flask',],
+)
