+++ web/web/templates/base.html
@@ -1,0 +1,8 @@
+<!doctype html>
+<title>TESS</title>
+{% if name %}
+  <h1>Hello {{ name }}!</h1>
+{% else %}
+  <h1>Hello, Worldzzzzz!</h1>
+{% endif %}
+
